Removed Parser for clean state to refactor
