+++ src/picounits/extensions/__init__.py
@@ -2,6 +2,6 @@
 # picounits/extensions/__init__.py
 
 from picounits.extensions.parser import Parser
-from picounits.extensions.loader import DynamicLoader
+from picounits.extensions.attribute_loader import DynamicLoader
 
 _, _ = Parser, DynamicLoader
