--- src/picounits/extensions/utilities/__init__.py
@@ -1,2 +1,0 @@
-# pylint: skip-file
-# picounits/extensions/utilities/__init__.py
